+++ src/interactive/widgets/mark.rs
@@ -31,7 +31,7 @@ pub enum MarkMode {
 
 pub type EntryMarkMap = BTreeMap<TreeIndex, EntryMark>;
 pub struct EntryMark {
-    pub size: u64,
+    pub size: u128,
     pub path: PathBuf,
     pub index: usize,
     pub num_errors_during_deletion: usize,
@@ -230,7 +230,7 @@ impl MarkPane {
         let title = format!(
             "Marked {} items ({}) ",
             marked.len(),
-            format.display(marked.iter().map(|(_k, v)| v.size).sum::<u64>())
+            format.display(marked.iter().map(|(_k, v)| v.size).sum::<u128>())
         );
         let selected = self.selected;
         let has_focus = self.has_focus;
